Move cache key to static method
